feat: add functionality to upload image to cloudinary

+++ server/src/modules/v1/posts/post.service.ts
@@ -1,7 +1,11 @@
+import cloudinary from '#src/cloudinary/setup';
 import authContextStorage from '#src/context/auth.context';
 import { Post } from '#src/generated/prisma';
 import { HTTPNotFoundError } from '#src/utils/errors/http.error';
-import { filterUndefinedValues } from '#src/utils/generic.util';
+import {
+    filterNullValues,
+    filterUndefinedValues,
+} from '#src/utils/generic.util';
 import prismaClient from '#src/utils/prisma-db/prisma-client.db';
 import fs from 'node:fs/promises';
 
@@ -11,12 +15,16 @@ export default class PostService {
         content,
         imageUrl,
     }: Pick<Post, 'title' | 'content' | 'imageUrl'>) => {
+        const { imagePublicId, imageSecureUrl } =
+            await this.uploadImageToCloudinary(imageUrl);
+
         // authorId from jwt token
         const newPost = await prismaClient.post.create({
             data: {
                 title,
                 content,
-                imageUrl,
+                imageUrl: imageSecureUrl,
+                imagePublicId,
                 authorId: authContextStorage.getContext('userId'),
             },
         });
@@ -49,16 +57,16 @@ export default class PostService {
         }: Pick<Post, 'title' | 'content' | 'imageUrl'>
     ) => {
         // check if new post imageUrl is sent, if so, delete existing image from uploads
-        if (imageUrl !== undefined) {
+        if (imageUrl !== undefined && imageUrl !== null) {
             const existingPost = await prismaClient.post.findUnique({
                 where: { id },
             });
-            const existingImageUrl = existingPost?.imageUrl;
-
-            if (existingImageUrl !== null && existingImageUrl !== undefined) {
-                await fs.unlink(existingImageUrl);
-            }
+            await this.deleteImageFromCloudinary(existingPost?.imagePublicId);
         }
+
+        // if new image is sent, save it to cloud
+        const { imagePublicId, imageSecureUrl } =
+            await this.uploadImageToCloudinary(imageUrl);
 
         // only update post when authorId matches userId from jwt, i.e only post owner can update post
         const updatedPost = await prismaClient.post.update({
@@ -66,7 +74,17 @@ export default class PostService {
                 id: id,
                 authorId: authContextStorage.getContext('userId'),
             },
-            data: { ...filterUndefinedValues({ title, content, imageUrl }) },
+            // IMPORTANT: filter out null values, to prevent overwriting image url and id when no new image is sent
+            data: {
+                ...filterNullValues({
+                    ...filterUndefinedValues({
+                        title,
+                        content,
+                        imageUrl: imageSecureUrl,
+                        imagePublicId,
+                    }),
+                }),
+            },
         });
 
         return updatedPost;
@@ -78,12 +96,40 @@ export default class PostService {
             where: { id, authorId: authContextStorage.getContext('userId') },
         });
 
-        const imageUrl = deletedPost.imageUrl;
-
-        if (imageUrl !== null) {
-            await fs.unlink(imageUrl);
-        }
+        await this.deleteImageFromCloudinary(deletedPost.imagePublicId);
 
         return deletedPost;
+    };
+
+    private uploadImageToCloudinary = async (imageUrl: string | null) => {
+        let imageSecureUrl = null;
+        let imagePublicId = null;
+
+        if (imageUrl !== null && imageUrl !== undefined) {
+            const cloudinaryResponse = await cloudinary.uploader.upload(
+                imageUrl,
+                {
+                    resource_type: 'image',
+                    // use_filename: true,
+                    // folder: 'node-express-api-posts',
+                }
+            );
+            imageSecureUrl = cloudinaryResponse.secure_url;
+            imagePublicId = cloudinaryResponse.public_id;
+
+            await fs.unlink(imageUrl);
+
+            return { imageSecureUrl, imagePublicId };
+        }
+
+        return { imageSecureUrl, imagePublicId };
+    };
+
+    private deleteImageFromCloudinary = async (
+        imagePublicId: string | null | undefined
+    ) => {
+        if (imagePublicId !== null && imagePublicId !== undefined) {
+            await cloudinary.uploader.destroy(imagePublicId);
+        }
     };
 }
